Add files via upload
Algorithm Lab-4

1.Given a sorted list of ‘n’ elements. You need to search a number ‘x’ in that
list. If the element ‘x’ is present in the list you have to return the index. If
it is not present return -1. Write an algorithm to implement the solution
to above problem using
a. Linear search.
b. Binary Search
c. K-ary search, where k= 3

2. An array of ‘n’ numbers has been given. You need to return a number in the array, which is neither maximum nor minimum.
a. Implement an O(n) algorithm to find the number, which is neither
maximum nor minimum when the array elements are not distinct.
b. In case, all the n-elements are distinct modify the implementation
to run the solution in O(1) time.

3. Let P1(x,y1),P2(x,y2),…,Pn(x,yn)P1(x,y1),P2(x,y2),…,Pn(x,yn) be a collection of nn distinct points lying on a vertical line LL. The value of x is stored in a variable, and y1,y2,…,yny1,y2,…,yn are stored in an array in decreasing order. Additionally, you are give